# swagger.yaml
openapi: 3.0.3
info:
  title: Todo Supabase Express API
  version: 1.0.0
  description: REST API for an Express todo application backed by Supabase.
servers:
  - url: http://localhost:3000
    description: Local development server
tags:
  - name: Health
    description: Server status endpoints
  - name: Todos
    description: Todo CRUD endpoints
paths:
  /api/health:
    get:
      tags:
        - Health
      summary: Check server health
      responses:
        "200":
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /api/todos:
    get:
      tags:
        - Todos
      summary: List todos
      description: Returns all todos ordered by newest first.
      responses:
        "200":
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags:
        - Todos
      summary: Create a todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoRequest"
      responses:
        "201":
          description: Todo created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/todos/{id}:
    parameters:
      - $ref: "#/components/parameters/TodoId"
    get:
      tags:
        - Todos
      summary: Get one todo
      responses:
        "200":
          description: Todo found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    put:
      tags:
        - Todos
      summary: Update a todo
      description: Updates one or more todo fields.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTodoRequest"
      responses:
        "200":
          description: Todo updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags:
        - Todos
      summary: Delete a todo
      responses:
        "204":
          description: Todo deleted
        "500":
          $ref: "#/components/responses/ServerError"
components:
  parameters:
    TodoId:
      name: id
      in: path
      required: true
      description: Todo UUID
      schema:
        type: string
        format: uuid
        example: 4a16d19c-2d5f-44e1-a64f-cfb511a55f44
  schemas:
    Todo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 4a16d19c-2d5f-44e1-a64f-cfb511a55f44
        title:
          type: string
          example: Learn Supabase
        description:
          type: string
          nullable: true
          example: Build CRUD routes with Express
        is_complete:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2026-09-03T12:00:00.000Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-09-03T12:00:00.000Z"
      required:
        - id
        - title
        - is_complete
        - created_at
        - updated_at
    CreateTodoRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          example: Learn Supabase
        description:
          type: string
          example: Build CRUD routes with Express
      required:
        - title
    UpdateTodoRequest:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          example: Learn Supabase deeply
        description:
          type: string
          example: Add validations and a browser UI
        is_complete:
          type: boolean
          example: true
    Error:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong.
      required:
        - error
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            missingTitle:
              value:
                error: Title is required.
            noFields:
              value:
                error: No valid fields provided.
    NotFound:
      description: Todo not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Todo not found.
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Something went wrong.